+++ vernac/debugHook.mli
@@ -10,22 +10,52 @@
 
 (** Ltac debugger interface; clients should register hooks to interact
    with their provided interface. *)
+
+(** The debugger receives requests by calling read_cmd to get Actions.
+ Each Action may return one or more responses as Answers by calling submit_answer.
+ Many of them return a single Answer or no Answer, but a single step may generate any
+ number of Outputs.
+
+ Debugger initialization has the following steps:
+ -> Coq sends Answer.Init
+ <- IDE sends zero or more initialization requests such as Action.UpdBpts
+ <- IDE sends Action.Configd
+
+ Stopping in the debugger generates Answer.Prompt and Answer.Goal messages,
+ at which point the IDE will typically call GetStack and GetVars.  When the
+ IDE sends with StepIn..Continue, the debugger will execute more code.  At
+ that point, Coq won't try to read more messages from the IDE until the
+ debugger stops again or exits.
+ *)
 module Action : sig
   type t =
-    | StepIn
-    | StepOver
-    | StepOut
-    | Continue
-    | Skip
-    | Interrupt
-    | Help
+    | StepIn    (* execute a single step in the tactic *)
+    | StepOver  (* execute steps until DB is back in the current stack frame *)
+    | StepOut   (* execute steps until DB exits current stack frame *)
+    | Continue  (* execute steps until a breakpoint or the debugger exits *)
+    | Skip      (* legacy: continue execution with no further debugging *)
+    | Interrupt (* exit the debugger *)
+    | Help      (* legacy: print help text *)
     | UpdBpts of ((string * int) * bool) list
-    | Configd
+                (* sets or clears breakpoints.  Values are:
+                   - absolute pathname of the the file
+                   - byte offset in the UTF-8 representation of the file
+                   - true to set, false to clear *)
+    | Configd   (* "config done" - indicates that the debugger has been
+                   configured, debugger does a Continue *)
+    | GetStack  (* request the call stack, returned as Answer.Stack *)
+    | GetVars of int
+                (* request the variables defined for stack frame N,
+                   returned as Answer.Vars.  0 is the topmost frame,
+                   followed by 1,2,3, ... *)
     | RunCnt of int
+                (* legacy: run for N steps *)
     | RunBreakpoint of string
+                (* legacy: run until an idtac prints the string *)
     | Command of string
-    | Failed
-    | Ignore (* do nothing, read another command *)
+                (* legacy: user-typed command to the debugger *)
+    | Failed    (* legacy: user command doesn't parse *)
+    | Ignore    (* internal: do nothing, read another command *)
 
   (* XXX: Should be moved to the clients *)
   val parse : string -> (t, string) result
@@ -33,10 +63,23 @@ end
 
 module Answer : sig
   type t =
-    | Prompt of Pp.t
-    | Goal of Pp.t
-    | Output of Pp.t
-    | Init
+    | Prompt of Pp.t (* output signalling the debugger has stopped
+                        Should be printed as a prompt for user input,
+                        e.g. in color without a newline at the end *)
+    | Goal of Pp.t   (* goal for the current proof state *)
+    | Output of Pp.t (* general output *)
+    | Init           (* signals initialization of the debugger *)
+    | Stack of (string * (string * int list) option) list
+                     (* The call stack, starting from TOS.
+                        Values are:
+                        - description of the frame
+                          (eg tactic name, line number, module)
+                        - absolute pathname of the file
+                        - array containing Loc.bp and Loc.ep of the
+                          corresponding code *)
+    | Vars of (string * Pp.t) list
+                     (* The variable values for the specified stack
+                        frame.  Values are variable name and variable value *)
 end
 
 module Intf : sig
@@ -52,9 +95,3 @@ module Intf : sig
   val set : t -> unit
   val get : unit -> t option
 end
-
-val forward_get_stack : (unit -> (string option * Loc.t option) list) ref
-val forward_get_vars : (int -> (string * Pp.t) list) ref
-
-val get_break : unit -> bool
-val set_break : bool -> unit